// app/(app)/wallets/[id]/types.ts
export interface Transaction {
  id: string;
  user_id: string;
  wallet_id: string | null;
  paylater_id: string | null;
  category_id: string | null;
  amount: number;
  type: "income" | "expense";
  description: string | null;
  transaction_date: string;
  receipt_url: string | null;
  created_at: string;
  wallets: { name: string; icon: string; color: string } | null;
  paylater_platforms: { name: string; color: string; icon: string } | null;
  categories: { name: string; icon: string; color: string } | null;
}